Асадчий Павел
гр. 2539

+++ RSS reader2/src/com/example/RSS_reader2/Updater.java
@@ -1,8 +1,12 @@
 package com.example.RSS_reader2;
 
 import android.app.IntentService;
+import android.app.ProgressDialog;
+import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.widget.Toast;
 
 import java.net.HttpURLConnection;
@@ -12,9 +16,8 @@ import java.util.ArrayList;
 
 public class Updater extends IntentService {
     ArticleDataBase articlesDataBase;
-    public static boolean isAlarm = false;
-    public static boolean isLast = false;
-
+    public static int goodArticle = 0;
+    public static ProgressDialog updater_dialog;
     public Updater() {
         super("some_name");
     }
@@ -23,13 +26,30 @@ public class Updater extends IntentService {
         super.onCreate();
     }
 
+    private boolean isNetworkAvailable() {
+        ConnectivityManager connectivityManager
+                = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+        return activeNetworkInfo != null;
+    }
+
     @Override
     protected void onHandleIntent(Intent intent) {
-        isAlarm = false;
-        isLast = false;
+        //updater_dialog = ProgressDialog.show(Updater.this, "Обновление данных. Пожалуйста, подождите...", null, true);
+        //updater_dialog.show();
+        if (isNetworkAvailable() == false) {
+            return;
+        }
+
+        goodArticle = 0;
         articlesDataBase = new ArticleDataBase(this);
         articlesDataBase.open();
         articlesDataBase.deleteALLArticles();
+        int t = -1;
+        while (t != 0) {
+            t = articlesDataBase.getCountArticles();
+        }
+
         Cursor cursor = articlesDataBase.sqLiteDatabase.query(ArticleDataBase.TABLE_FEED_NAME, new String[] {
                 ArticleDataBase.KEY_ID_FEED, ArticleDataBase.KEY_FEED_NAME, ArticleDataBase.KEY_FEED},
                 null, // The columns for the WHERE clause
@@ -40,16 +60,21 @@ public class Updater extends IntentService {
         );
 
         while (cursor.moveToNext()) {
-            // GET COLUMN INDICES + VALUES OF THOSE COLUMNS
+            String feed = cursor.getString(cursor.getColumnIndex(ArticleDataBase.KEY_FEED));
+            if (isCorrect(feed)) {
+                ++goodArticle;
+            }
+        }
+
+        cursor.moveToFirst();
+        while (cursor.moveToNext()) {
             String feed = cursor.getString(cursor.getColumnIndex(ArticleDataBase.KEY_FEED));
             if (isCorrect(feed)) {
                 Intent intentForRss = new Intent(this, MainIntentWork.class);
                 intentForRss.putExtra(ArticleActivity.KEY_FOR_INTENT, feed);
                 startService(intentForRss);
-                isAlarm = true;
             }
         }
-        isLast = true;
     }
 
     private boolean isCorrect(String feed) {
